Remove "run as root" feature
Decided that this feature should never have been added. It does not fit
in with the use cases targeted by this project. It is also not safe,
there are better ways to accomplish the goal of taking a backup of /etc
using kup.
Basically: if you are an advanced user, someone who backs up /etc, then
you can also set up an admin group, add yourself to that one and modify
files previously only readable by root to be readable by that group.
Or something like that.

bup was not written to be run as suid root.