Make filedigger significantly faster for large directories
By not checking file sizes for all versions until asked for the info by the versionlistmodel.

Also, quite important! Fix wrong "modified" time stamps for files. Bug in generateSubnodes(), was not reading metadata from the metadata stream for all files, needs to be read unconditionally so that later files will match up with the stream.