Bug fix for detecting mounts on a directory
The fs plan executor did not detect when something got mounted on an existing directory. It only worked with udisk mounting since then a directory is created for use as mount point and kup detected the creation of the directory (and mount happened within 5 seconds after directory creation).
Now it also detect mounts on existing directories by monitoring /proc/mounts for changes, reads that file to find out if the watched directory turns up as a mount point.

Fixes #12.