Bug fix: correct showing of hidden folders in source selection
The bug was that an included or excluded folder which itself was not hidden but which was a subdir of a hidden dir would cause the treeview to go all collapsed and not working if the "show hidden folders" option was not set true.
Also add a small feature; when enabling "show hidden folders" then run the expand logic again to expand any folders that could now have become visible.