Change policy for rsync destination folder
New policy:
1: If source selection only includes one folder, put the contents of that
   folder directly under the destination folder.
2: If backing up two or more folder then look for a naming conflict
   between the source folder names. If there's a conflict then put full
   path hierarchy under the destination folder.
3: If no conflict then put the source folders in the destination folder
   but not any parents of the source folders.

This should minimize the depth of folders, avoiding a destination folder
that just contains one folder which just contains one folder which contains
the backed up stuff. Rsync would merge the contents of two folders with
the same name, so to avoid that we look for any name conflict.

If the user ever deselects a the source folder of the backup plan then
the next time kup runs rsync the files that were previously copied will
not be removed. That was true before and is still true. This change will
make that situation happen more often, now it can happen not only when
deselecting a folder but also when going from one to two selected folders.
But not when going from two to more folders... except if that gives a
naming conflict.

Fixes #29.