Add ability to exclude files and folders based on patterns
For now it's just a simple selection of text file.
Idea is that it is better to have some ability to do this than wait for
a fancy rule GUI that may never happen because nobody has time to create
such a thing.
Bup and rsync having different ways of specifying rules also makes such
a fancy solution even more difficult.