Add detection of removed source folders
Reason for aborting backup save job is that I think the most likely
scenario triggering this is that a folder has been renamed. So we don't
want to automatically remove the missing folder from backup plan and carry
on saving remaining stuff. Instead, refuse to do anything until user opens
settings and confirm again what they want to include. To make that easy
we now offer a button to take user directly to sources page of settings
module.