Go back to using `kup` as the icon
Using `preferences-system-backup` is not really semantically correct as
that is a generic icon. Apps should be using their own icon.

This will result in no visual changes for now since `kup` is symlinked
to that other icon anyway in Breeze, but this may change at some point
in the future, and we want to support that.