Remove connection error text from data source
The error message displayed to user was not translated and regardless of
what the problem was, would not help the user figure out how to solve it.
If there even was a problem. Because it could also be displayed if the
plasmoid was forced to be displayed even when kup backups was not enabled.

BUG: 444704