Add a fileitemaction plugin
This adds a plugin for use by file managers to give an additional action
for selected files and folders. The action is to offer restoring
previous version of files or missing files for folders. The action will
only be provided if the selected file is included in a versioned backup
plan and the saved backups are currently available in the system.

Kup daemon gets a new d-bus exported function to get the path of a bup
repository for a given file path. It returns an empty string for files
where no backup is available.

BUG: 499455